+++ cpsycle/ui/src/uiapp.h
@@ -71,7 +71,7 @@ void psy_ui_app_setdefaultfont(psy_ui_App*, psy_ui_Font*);
 void psy_ui_app_lighttheme(psy_ui_App*);
 void psy_ui_app_darktheme(psy_ui_App*);
 bool psy_ui_app_hasdarktheme(const psy_ui_App*);
-const psy_ui_Style* psy_ui_app_style(const psy_ui_App*, int styletype);
+const psy_ui_Style* psy_ui_app_style(const psy_ui_App*, uintptr_t styletype);
 void psy_ui_app_updatesyles(psy_ui_App*);
 void psy_ui_app_sethover(psy_ui_App*, struct psy_ui_Component* hover);
